requirements.txt
python-telegram-bot>=13.7
supabase>=0.7.1
python-dotenv>=0.19.0
psycopg2-binary>=2.9.1
pytz>=2023.3
requests>=2.27.1
uuid
